Separating auth config properties
