+++ main/model/__init__.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 
 from .base import Base
+from .config_auth import ConfigAuth
 from .config import Config
 from .user import User
